Rework items to the new 'everything in module' format
